requirements.txt
a2wsgi==1.10.10
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.9.3
apache-airflow-core==3.0.6
apache-airflow-providers-amazon==9.12.0
apache-airflow-providers-common-compat==1.7.3
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apache-airflow-providers-standard==1.6.0
apache-airflow-task-sdk==1.0.6
apispec==6.6.1
argcomplete==3.4.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.13.5
blinker==1.8.2
boto3==1.40.22
botocore==1.40.22
cachelib==0.9.0
cadwyn==5.4.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.3
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.3
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docutils==0.16
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.59.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==0.10.0
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy-object-proxy==1.10.0
libcst==1.8.2
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.4.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
minio==7.2.16
more-itertools==10.3.0
msgspec==0.19.0
multidict==6.0.5
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.1
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
pendulum==3.0.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.5.0
ply==3.11
prison==0.2.1
prompt_toolkit==3.0.51
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
PyAthena==3.17.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.2.3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
pywin32==311
PyYAML==6.0.1
pyzmq==27.0.2
redshift-connector==2.1.8
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.19.0
s3transfer==0.13.1
sagemaker_studio==1.0.21
scramp==1.4.6
seaborn==0.13.2
sentry-sdk==2.35.2
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.47.3
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.2
tornado==6.5.2
traitlets==5.14.3
typer==0.17.3
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==1.26.19
uuid6==2025.0.1
uvicorn==0.35.0
watchfiles==1.1.0
watchtower==3.4.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
xlrd==2.0.2
yarl==1.9.4
zipp==3.19.2
